chore(refactor): Split out occurrence-checking code
